src/demo/requirements_demo.txt
gradio>=4.0.0
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
numpy>=1.21.0
transformers>=4.20.0
faiss-cpu>=1.7.0
tqdm>=4.60.0 